triton-florence/requirements.txt
vllm==0.2.5
torch>=2.0.0
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.24.0

